feat: Add slider functionality to index.html
This commit adds JavaScript code to enable slider functionality on the index.html page. The slider allows users to navigate through a collection of project cards. The previous and next buttons update the current index and translate the slider wrapper accordingly. This enhances the user experience by providing an interactive way to view different projects.